+++ DirectXMesh/DirectXMeshVBWriter.cpp
@@ -163,7 +163,7 @@ HRESULT VBWriter::Impl::Initialize(const InputElementDesc* vbDecl, size_t nDecl)
     ComputeInputLayout(reinterpret_cast<const D3D11_INPUT_ELEMENT_DESC*>(vbDecl), nDecl, offsets, mDefaultStrides);
 #endif
 
-    for (uint32_t j = 0; j < nDecl; ++j)
+    for (size_t j = 0; j < nDecl; ++j)
     {
         if (vbDecl[j].InputSlotClass == PER_INSTANCE_DATA)
         {
@@ -176,18 +176,18 @@ HRESULT VBWriter::Impl::Initialize(const InputElementDesc* vbDecl, size_t nDecl)
 
         mInputDesc[j].AlignedByteOffset = offsets[j];
 
-        auto decl = SemanticMap::value_type(vbDecl[j].SemanticName, j);
+        auto decl = SemanticMap::value_type(vbDecl[j].SemanticName, static_cast<uint32_t>(j));
         mSemantics.insert(decl);
 
         // Add common aliases
         if (_stricmp(vbDecl[j].SemanticName, "POSITION") == 0)
         {
-            auto decl2 = SemanticMap::value_type("SV_Position", j);
+            auto decl2 = SemanticMap::value_type("SV_Position", static_cast<uint32_t>(j));
             mSemantics.insert(decl2);
         }
         else if (_stricmp(vbDecl[j].SemanticName, "SV_Position") == 0)
         {
-            auto decl2 = SemanticMap::value_type("POSITION", j);
+            auto decl2 = SemanticMap::value_type("POSITION", static_cast<uint32_t>(j));
             mSemantics.insert(decl2);
         }
     }
